requirements.txt
anyio==4.6.2
asgiref==3.8.1
attrs==24.3.0
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
diff-match-patch==20241021
distlib==0.3.8
dj-database-url==2.2.0
Django==5.0.7
django-countries==7.6.1
django-debug-toolbar==4.4.6
django-environ==0.11.2
django-extensions==3.2.3
django-heroku==0.3.1
django-import-export==4.3.5
django-unfold==0.49.1
django-widget-tweaks==1.5.0
django_paystack==1.0.0
djangorestframework==3.15.1
feedparser==6.0.11
git-filter-repo==2.45.0
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
outcome==1.3.0.post0
packaging==24.0
pillow==10.4.0
platformdirs==4.2.2
psycopg2-binary
pycparser==2.22
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
selenium==4.27.1
setuptools==75.8.0
sgmllib3k==1.0.0
simplejson==3.19.3
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sqlparse==0.5.1
sweetify==2.3.1
tablib==3.8.0
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.2
virtualenv==20.26.3
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
whitenoise==6.6.0
wsproto==1.2.0
